Fix bug in setting value for terminal nodes
The bug caused the terminal value to be always set to -1.  Also noting that it's
not possible for the terminal value to be 1, since the game either ends in a
draw or with the side to move being mated.  So essentially the bug meant that we
were incorrectly assigning a value to nodes that terminated in a draw.