Add SearchInfo struct and update AgentZero interface
Store search parameters in SearchInfo struct.  This is in preparation of adding
search output when running in UCI mode.

The uci_loop interface has been switched from Agent* to ZeroAgent*, since it
accesses the SearchInfo data, whcih isn't currently part of the base class.
Suspect that this is appropriate, as the search and options will be specific to
the agent.  But it could be generalized if needed by making SearchInfo part of
the Agent base class, possibly using a SearchInfo base class as well.