Modify network policy stack architecture per AZ supp materials
Instead of flattening and using a linear layer, we use two convolutions.  One is
a copy of hte block from the body, and the last convolution outputs the 73
planes for the move policies.  We do not perform the softmax in the network, so
this will need to be done in the training code.  This also facilitates applying
a softmax temperature in search.

Interestingly, this reduces the number of parameters by more than a factor of 3.
Inference throughput seems to be about the same with laptop CPU.