Initial WIP to switch to ONNX for nerual net inference
Currently the library part of the code compiles, but there are some errors with
the executables.  There are still placeholders where things will need to be
updated, but just wanted to get it to compile.

The idea is to use the ONNX runtime instead of libtorch for inference.  Some
testing suggests that this might be 2-3x faster.  The library also seems to be
more lightweight.  The main difference in the implementation is swapping out
Torch tensors for a lightweight custom tensor wrapper.  The ONNX runtime
basically works directly with raw arrays, so I have to manage the dimensions,
indexing, and ownership of the memory.