Implement experience serialization with new tensors
The main difference is that with torch, we used a library function to
concatenate everything together before serializaing, whereas here we leave them
as a vector of tensors and concatenate during write.  We also have special case
functions for handling the reward data, which are just stored in a vector of
scalars.