Update pytorch scripts to export ONNX models
Add a new option to conv_4x64.py that allows supplying an input model network.
This can be used to convert an existing model to ONNX.

One minor change for both files was to set requires_grad=True on the sample
input.  AFAICT this doesn't make a difference, but it is used in the online
examples that I have found.