Add "simple" time manager based on LC0
This time manager uses a base percentage of time left and then adjusts for both move
count and increment fraction.  From informal testing on short increment time
controls (e.g., 2min/1sec), it seems to do a much better job of managing the time near
the end, as compared to the AlphaZeroTimeManager.

Fo now, the SimpleTimeManager replaces AlphaZeroTimeManager as the default in UCI mode
in main.cpp.  An option for selecting between them could be added in the future.