Make Experience.complete_episode side agnostic
The sides are now stored as part of the experience data.  This makes it possible for
complete_episode to be called with the reward for the white side, and then to fill all
of the reward data appropriately based on the actual side.

This will make it possible to use a single agent and experience collector for both sides
in selfplay, enabling tree use across moves.