Remove code for tree reuse
Discovered that LC0 doesn't enable tree reuse for selfplay, and selfplay was really the
only case I was covering.  Since selfplay includes the use of noise at the root, it
makes sense that tree reuse could interfere with that.  Instead of adding a flag to
disable it, decided to cut it out for simplicity, since I didn't have use for it outside
of selfplay (wasn't even checking deep enough for it to come into play in match play
with different agents).