Initial infrastructure for neural network cache
This implements a map/set that parallels a FIFO queue, which can serve as a cache with a
maximum size.  As items are added, we evict old items (based on the queue) when the
specified size is exceeded.

Some helper functions have been added (adopted from LC0) for concatenating hashes.  This
is used to augment the raw board hash to include information about repetition and
no-progress counts, which is part of the NN encoding.

Some debug information is printed out about to show the number of cache hits, but no
real caching of NN results is done yet.  Currently the raw output of the NN for the
prior is a large sparse array, which gets decoded.  Ideally, the cache will store the
compact, decoded value instead of the large raw output.