Store and access network inference results from cache
Define NetworkOutput struct that contains the prior and value data.  This is used as the
value type in the cache map.  The cached inference eval function now returns a reference
to a NetworkOutput in the cache.  If necessary, it first comptues the output and stores
it in the cache.  By returning a reference, we aviod copying the data for move priors
back out of the cache.