Avoid always copying the CachedInferenceModel output
Previously was declaring the output variable as "auto", which ends up being a value
type, not a reference (even though the function output is a const reference).  This
meant that the result always got copied.  This is why the compiler didn't complain about
modifying the local variable in the call to add_noise_to_priors.  This is addressed by
declaring the output variable as "auto&" which makes it a (const) reference to the data
in cache.  As such, we can't (and don't want to) modify it when adding noise.  To work
around this, a local pointer is declared and assigned to the address of the either the
cache reference (when not adding noise) or a local copy (when adding noise).