Initial work to add board transforms
Following LC0, we can transform the board into a canonical representation before
building the neural network input.  This commit adds a choose_transform that decides
which transforms can be applied.  It follows LC0, but leaves out some of the more
detailed handling when a tranposition doesn't affect the king position.

No actual changes to the encoding are done yet.  For now, there is a temporary counter
to check the frequency that such transformations would be applied.  From simple testing,
it seems to be about 20-40% for an untrained network, and less for a trained network.