Add benchmark_onnx.py
Add script to benchmark inference using ONNX from python.  Also add pip install of
onnxruntime to the Dockerfile (it's only needed for running the model from Python, but
it doesn't seem to significantly increase the container size).

The benchmark script includes a batch option.  Using something other than 1 requires
exporting the model with dynamic axes (the commented-out options for this are included
in conv_4x64.py).