Add command line options for specifying encoding version
Add command options to main.cpp, selfplay.cpp, and conv_4x64.py.  This should be
sufficient for running training with this option enabled.  Marked the C++ options as
todo, because should be able to parse the input shape from the ONNX model and determine
the encoding dynamically based on that.