Add sticky_num_rounds flag to override UCI go infinite
This flag is set if the number of rounds is specified on the command line.  In this
case, the setting for number of rounds will still be used even with UCI "go infinite".
This makes it possible to use fixed playouts in cutechess tournaments (via the infinite
time control setting).