Switch from 32 to 4 output filters in value stack
Reduce from 32 filters (LC0 settings) to 4 filters in the convolution layer in the
output stack.  My thinking is that this is already more filters than was used in AZ (1
filter), and 32 might be overkill relative to the size of the rest of my network.  Due
to the linear layer, the number of filters has a big effect on the total number of
parameters, especially relative to the rest of network.  Some quick testing shows that
although the effect on number of parameters is large, there is only a small effect on
throughput.

Note that going all the way down to 1 filter probably doesn't make sense, given that the
linear layer would then be increasing dimension from 64 to 128.  With 4 filters, we
still get some dimension reduction in the linear layer.